requirements.txt
docker_py==1.10.6
packaging==20.1
PyGithub==1.58.1
pytest==5.3.5
PyYAML==6.0
requests==2.28.2
slack_sdk==3.20.2
